Split node implemention logic into CLI and API versions.
